Made ultrasonic localizer class

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utils/Pose.kt
@@ -1,9 +1,9 @@
 package org.firstinspires.ftc.teamcode.utils
 
-data class Pose(val x: Float, val y: Float, val rotation: Float) {
+data class Pose(val x: Double, val y: Double, val rotation: Double) {
 
-    constructor(): this(0f, 0f, 0f)
-    constructor(v2: Vector2, rotation: Float): this(v2.x, v2.y, rotation)
+    constructor(): this(0.0, 0.0, 0.0)
+    constructor(v2: Vector2, rotation: Double): this(v2.x, v2.y, rotation)
 
     fun extractPosition(): Vector2 {
         return Vector2(x, y)
