+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utils/Vector2.kt
@@ -2,10 +2,10 @@ package org.firstinspires.ftc.teamcode.utils;
 
 import kotlin.math.sqrt
 
-data class Vector2(val x: Float, val y: Float) {
-    constructor(): this(0f, 0f)
+data class Vector2(val x: Double, val y: Double) {
+    constructor(): this(0.0, 0.0)
 
-    fun magnitude(): Float {
+    fun magnitude(): Double {
         return sqrt(x * x + y * y)
     }
 }
